+++ npm/grep/cypress/support/e2e.ts
@@ -1,7 +1,7 @@
 // @ts-check
 /// <reference types="cypress" />
 
-import { register } from '@cypress/grep'
+import { register } from '../../src/index'
 
 // register the grep feature
 // https://github.com/cypress-io/cypress-grep
